// src/app/shared/interfaces/hwm.ts
export interface hwm {
    hwm_id: number;
    waterbody?: string;
    site_id?: number;
    event_id?: number;
    eventName?: string;
    hwm_type_id: number;
    hwmType?: string;
    hwm_quality_id: number;
    hwmQuality?: string;
    hwm_locationdescription?: string;
    latitude_dd?: number;
    longitude_dd?: number;
    survey_date?: Date;
    elev_ft?: number;
    vdatum_id?: number;
    vDatum?:string;
    vcollect_method_id?: number;
    vCollectMethod?: string;
    bank?: string;
    approval_id?: number;
    marker_id?: number;
    markerName?: string;
    height_above_gnd?: number;
    hcollect_method_id?: number;
    hCollectMethod?: string;
    peak_summary_id?: number;
    hwm_notes?: string;
    hwm_environment?: string;
    flag_date?: Date;
    stillwater?: number;
    hdatum_id?: number;
    hDatum?: string;
    flag_member_id?: number;
    flagMember?: string;
    survey_member_id?: number;
    surveyMember?: string;
    uncertainty?: number;
    hwm_uncertainty?: number;
    hwm_label?: string;
}